Separated out hte "tab panel" into page-specific components for readability.
