+++ src/Components/HomeTabPanel.js
@@ -1,0 +1,68 @@
+// Frontend UI 
+import React, {useContext, 
+    useEffect, 
+    useRef,
+    useState,} from "react";
+import { useNavigate } from "react-router";
+import PropTypes from 'prop-types';
+import {Container,
+    Box, 
+    Tabs, 
+    Tab, 
+    Typography} from "@material-ui/core";
+import TabPanel from "./TabPanel";
+
+// Plotting
+import Highcharts from 'highcharts';
+import HighchartsReact from "highcharts-react-official";
+import {AgGridReact} from 'ag-grid-react';
+import 'ag-grid-community/dist/styles/ag-grid.css';
+import 'ag-grid-community/dist/styles/ag-theme-alpine.css';
+
+// Contexts (TASK: REMOVE IF NOT NECESSARY)
+import OrganismContext from "../Contexts/OrganismContext";
+import ClusterContext from "../Contexts/ClusterContext";
+
+
+function HomeTabPanel(props) {
+    const [value, setValue] = React.useState(0);
+
+    const handleChange = (event, newValue) => {
+      setValue(newValue);
+    };
+  
+    return (
+      <Box sx={{ width: '100%' }}>
+  
+        <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
+          <Tabs value={value} onChange={handleChange} aria-label="basic tabs example">
+            <Tab label="About" {...a11yProps(0)} />
+            <Tab label="Help" {...a11yProps(1)} />
+          </Tabs>
+        </Box>
+  
+        <TabPanel value={value} index={0}>
+          Description of the software tool 
+        </TabPanel>
+        <TabPanel value={value} index={1}>
+          Instructions for how to use it 
+        </TabPanel>
+  
+      </Box>
+    );
+}
+
+HomeTabPanel.propTypes = {
+    children: PropTypes.node,
+    index: PropTypes.number.isRequired,
+    value: PropTypes.number.isRequired,
+};
+  
+function a11yProps(index) {
+    return {
+      id: `simple-tab-${index}`,
+      'aria-controls': `simple-tabpanel-${index}`,
+    };
+}
+
+export default HomeTabPanel;
